requirements.txt
aws-cdk-lib==2.113.0
constructs>=10.0.0,<11.0.0

boto3~=1.33.8
botocore~=1.33.8